Finish folder picker & rewrite program to support unicode paths

+++ restore.cpp
@@ -7,7 +7,7 @@
 restore::restore()
 {
 	inipath iniPath;
-	std::string targetPath = iniPath.getPath();
+	std::wstring targetPath = iniPath.getPath();
 	if (!std::filesystem::exists(iniPath.getBackupDir()))
 	{
 		MessageBoxA(NULL, "No backups found for the specified path.", "Backup failed!", MB_OK | MB_ICONERROR);
@@ -19,11 +19,21 @@ restore::restore()
 		if (fs.path().extension() == ".ini" || fs.path().extension() == ".INI")
 		{
 			// Modify the file paths as if they reside in the target path for easy file replacement
-			std::string destination = fs.path().string();
+			std::wstring destination = fs.path().wstring();
 			destination.erase(0, iniPath.getBackupDir().size());
 			destination.insert(0, iniPath.getPath());
-			std::filesystem::remove(destination);
-			std::filesystem::copy(fs.path().string(), destination);
+			if (std::filesystem::exists(destination))
+			{
+				std::filesystem::remove(destination);
+				std::filesystem::copy(fs.path().wstring(), destination);
+			}
+			else
+			{
+				std::wstring parentFolder = destination;
+				parentFolder.erase(parentFolder.rfind(L"\\"), parentFolder.length());
+				std::filesystem::create_directory(parentFolder);
+				std::filesystem::copy(fs.path().wstring(), destination);
+			}
 		}
 	}
 }
